+++ R/print.parameters_model.R
@@ -87,6 +87,11 @@
 #'   labels will be used as parameters names. The latter only works for "labelled"
 #'   data, i.e. if the data used to fit the model had `"label"` and `"labels"`
 #'   attributes. See also section _Global Options to Customize Messages when Printing_.
+#' @param add_reference Logical, if `TRUE`, the reference level of factors will
+#'   be added to the parameters table. This is only relevant for models with
+#'   categorical predictors. The coefficient for the reference level is always
+#'   `0` (except when `exponentiate = TRUE`, then the coefficient will be `1`),
+#'   so this is just for completeness.
 #' @inheritParams insight::format_table
 #' @inheritParams compare_parameters
 #'
@@ -233,6 +238,7 @@ print.parameters_model <- function(x,
                                    groups = NULL,
                                    column_width = NULL,
                                    ci_brackets = c("[", "]"),
+                                   add_reference = FALSE,
                                    ...) {
   # save original input
   orig_x <- x
@@ -282,6 +288,7 @@ print.parameters_model <- function(x,
     ci_brackets = ci_brackets,
     format = "text",
     groups = groups,
+    add_reference = add_reference,
     ...
   )
 
@@ -378,6 +385,7 @@ print.parameters_random <- function(x, digits = 2, ...) {
                         ci_brackets = TRUE,
                         format = "text",
                         group = NULL,
+                        add_reference = FALSE,
                         ...) {
   format(
     x,
@@ -392,6 +400,7 @@ print.parameters_random <- function(x, digits = 2, ...) {
     zap_small = zap_small,
     format = format,
     group = group,
+    add_reference = add_reference,
     ...
   )
 }