+++ R/utils_format.R
@@ -280,6 +280,7 @@
                                              format = NULL,
                                              coef_name = NULL,
                                              zap_small = FALSE,
+                                             add_reference = FALSE,
                                              ...) {
   # default brackets are parenthesis for HTML / MD
   if ((is.null(ci_brackets) || isTRUE(ci_brackets)) && (identical(format, "html") || identical(format, "markdown"))) {
@@ -308,6 +309,11 @@
     x$Level <- NULL
   }
 
+  # add the coefficient for the base-(reference)-level of factors?
+  if (add_reference) {
+    x <- .add_reference_level(x)
+  }
+
   insight::format_table(
     x,
     pretty_names = pretty_names,
@@ -319,6 +325,86 @@
     zap_small = zap_small,
     ...
   )
+}
+
+
+.add_reference_level <- function(params) {
+  # check if we have a model object, else return parameter table
+  model <- .get_object(params)
+  if (is.null(model)) {
+    params
+  }
+
+  # check if we have model data, else return parameter table
+  model_data <- insight::get_data(model)
+  if (is.null(model_data)) {
+    params
+  }
+
+  # find factors and factor levels and check if we have any factors in the data
+  factors <- .find_factor_levels(model_data)
+  if (!length(factors)) {
+    params
+  }
+
+  # we need some more information about prettified labels etc.
+  pretty_names <- attributes(params)$pretty_names
+  coef_name <- attributes(params)$coefficient_name
+  if (is.null(coef_name)) {
+    coef_name <- "Coefficient"
+  }
+  zi_coef_name <- attributes(params)$zi_coefficient_name
+  if (is.null(zi_coef_name)) {
+    zi_coef_name <- "Coefficient"
+  }
+
+  # copy object, so we save original data
+  out <- params
+
+  # if we use "keep" or "drop", we have less parameters in our data frame,
+  # so we need to make sure we only have those pretty_names, which names match
+  # the parameters in the data frame
+  pretty_names <- pretty_names[names(pretty_names) %in% params$Parameter]
+
+  # iterate all factors in the data and check if any factor was used in the model
+  for (fn in names(factors)) {
+    f <- factors[[fn]]
+    # "f" contains all combinations of factor name and levels from the data,
+    # which we can match with the names of the pretty_names vector
+    found <- which(names(pretty_names) %in% f)
+    # if we have a match, we add the reference level to the pretty_names vector
+    if (length(found)) {
+      # the reference level is *not* in the pretty names yet
+      reference_level <- f[!f %in% names(pretty_names)]
+      # create a pretty level for the reference category
+      pretty_level <- paste0(fn, " [", sub(fn, "", reference_level, fixed = TRUE), "]")
+      # insert new pretty level at the correct position in "pretty_names"
+      pretty_names <- .insert_element_at(
+        pretty_names,
+        stats::setNames(pretty_level, reference_level),
+        min(found)
+      )
+      # now we need to update the data as well (i.e. the parameters table)
+      row_data <- data.frame(
+        Parameter = reference_level,
+        Coefficient = as.numeric(attributes(params)$exponentiate),
+        stringsAsFactors = FALSE
+      )
+      # coefficient name can also be "Odds Ratio" etc., so make sure we
+      # have the correct column name in the data row we want to insert
+      if (coef_name %in% colnames(out)) {
+        colnames(row_data)[2] <- coef_name
+      } else if (zi_coef_name %in% colnames(out)) {
+        colnames(row_data)[2] <- zi_coef_name
+      }
+      out <- .insert_row_at(out, row_data, min(found))
+    }
+  }
+  # update pretty_names attribute
+  attr(out, "pretty_names") <- pretty_names
+  attr(out, "pretty_labels") <- pretty_names
+
+  out
 }
 
 
@@ -765,6 +851,7 @@
                                                 ci_width = "auto",
                                                 ci_brackets = TRUE,
                                                 zap_small = FALSE,
+                                                add_reference = FALSE,
                                                 ...) {
   final_table <- list()
 
@@ -962,6 +1049,11 @@
       sample_rows <- round(c(1, stats::quantile(seq_len(n_rows), seq_len(row_steps - 2) / row_steps), n_rows))
       tables[[type]] <- tables[[type]][sample_rows, ]
       tables[[type]][[1]] <- insight::format_value(tables[[type]][[1]], digits = digits, protect_integers = TRUE)
+    }
+
+    # add the coefficient for the base-(reference)-level of factors?
+    if (add_reference) {
+      tables[[type]] <- .add_reference_level(tables[[type]])
     }
 
     formatted_table <- insight::format_table(