+++ NEWS.md
@@ -8,6 +8,10 @@
 * `model_parameters()` for objects from package *marginaleffects*  now also
   accepts the `exponentiate` argument.
 
+* The `print()`, `print_html()`, `print_md()` and `format()` methods for
+  `model_parameters()` get an `add_reference` argument, to add the reference
+  category of categorical predictors to the parameters table.
+
 ## Bug fixes
 
 * Fixed issue with wrong calculation of test-statistic and p-values in